Add service account for UHC clusters service
The UHC clusters service needs to run with an specific service account
in order to give it permissions to create the configuration map that it
will use to elect a leader that will receive an process events.

This service account can't be created in the template of the service
because _saasherder_ filters them out.

The service account needs to be linked to the _quay.io_ pull secret in
order to pull the private images of the containers.

Related: https://jira.coreos.com/browse/SDA-517
Signed-off-by: Juan Hernandez <jhernand@redhat.com>