resources/datahub-psi/dh-stage-prometheus-lts/dashboards/grafana-dashboard-ocp-overview.configmap.yaml: Add degraded operator reasons
Because it's good to have some idea why the operators are failing
without having to drill down into the single-operator dashboard [1].

Also shift the legend off to the right.  This is a wide panel, so the
resulting reduction in chart width doesn't hurt readability as much as
the reduction in chart height from having the legend underneath.

I'd initially proposed adding topk(5, ...) around the whole query, in
the hopes that it might make the calculation faster and with the
expectation that lower-likelyhood issues are less likely to be
critical failures (or at least were less likely to get investigated
and addressed).  But Clayton asked me to drop that [2], so I have.

[1]: https://telemeter-lts-dashboards.datahub.redhat.com/d/uzoviEnZk/pre-production-cluster-operators?orgId=1
[2]: https://gitlab.cee.redhat.com/service/app-interface/merge_requests/1367#note_793627