Add `SANDBOX` environment to OCM merge request jobs
We are working on a proof of concept to add tests that require a running
database to the OCM projects. Those tests require a cluster where a
temporary project can be created to deploy that database. To do so we
would like to create a bare metal cluster that we have created in the
lab in Tel Aviv. The credentials and the proxy needed to use that
cluster have been added to the `config` and `proxy` keys of a new secret
in Vault. This patch adds those keys as environment variables to the OCM
merge request check jobs.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>