Add logs publisher to OCM `pr_check` jobs
The logs of the OCM `pr_check` jobs are very verbose now that we run the
integration tests before merging. It isn't convenient to have everything
in the same file. To improve that this patch modifies the job template
so that jobs will publish the files inside the `logs` directory. Later
patches will change the projects to put there the relevant log files.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>