Move OCM sandbox token and url to separate Vault keys
Currently the configuration of the cluster used for OCM tests (a.k.a.
the _sandbox_) is stored in a single `config` key in Vault, which is
then mapped to a single `SANDBOX_CONFIG` environment variable in
Jenkins.

This patch changes that so that the token and the URL of the cluster are
in separate `token` and `url` keys mapped to separate `SANDBOX_TOKEN`
and `SANDBOX_URL` environment variables in Jenkins.

This should simplify updating this information in the future.

Related: https://jira.coreos.com/browse/APPSRE-1096
Signed-off-by: Juan Hernandez <jhernand@redhat.com>