Fix OCM controller manager down alert
Currently we are checking if there is no controller manager running, but
this doesn't detect situations where one controller manager is down but
others are running. This patch fixes that so that the alerts will be
triggered when any of the controller managers goes down.

Related: https://jira.coreos.com/browse/SDA-1653
Signed-off-by: Juan Hernandez <jhernand@redhat.com>