Connect router directly to backend services
A recent change in the Telemetry client exposed the fact that the API
gateway is not ready to process a very high volume of requests. To
improve that this patch changes the configuration of the router used in
the production environment so that it connects directly to the backend
services.

This can't be done in the integration or staging environments because
the API gateway is the component that checks that the user has a
redhat.com e-mail address, and there is no way to do that with the
router directly.

Access to the Swagger UI and to other paths that don't belong to any
specific backend will still go through the API gateway.

https://jira.coreos.com/browse/SDA-1697
Signed-off-by: Juan Hernandez <jhernand@redhat.com>