[stage]Fix number of clusters reporting OLM metrics figure
The OLM Managed Operators dashboard is incorrectly reporting
the number of clusters which have OLM metrics enabled in them.
This PR fixes that number.

Incorrect assumption before: All clusters have subscription
for the packageserver, which was being scraped to count the
total number of clusters that have OLM installed in them.
However, the subscription for the packageserver was removed
from 4.2 onwards, and therefore only 4.1 clusters with OLM
installed in them were being counted.

Solution: Take into account the 4.1 clusters that are emitting
metrics from packageserver subscription, and ALL >=4.2 clusters
to count the total number of clusters which are capable of emitting
OLM metrics.