resources/datahub-psi/dh-*-prometheus-lts/dashboards/grafana-dashboard-ocp-cluster-version-upgrades: Backslash to bracket expressions
With the previous values, selecting, for example, 4.2 for the version
would result in:

  version=~"4\.2\.\d+"

in the Grafana queries and:

  var-version=4\.2\.\d%2B

as the URI query parameter.  But we need an additional round of
escaping, because Grafana consumes one level locally and we still want
backslashes to get through to Prometheus.  The old values (with one
level too little escaping) were from 52a5dacc62 (datahub-psi: Improve
the upgrade dashboard queries, 2019-12-17, !2475).

With this commit, I pivot from backslash escapes to bracket
expressions.  That avoids the JSON-sensitive characters and means we
no longer have to worry about how many backslash expansions will
happen between the config map tracked in this repository and the
eventual PromQL engine.