resources/datahub-psi/dh-*-prometheus-lts/dashboards/grafana-dashboard-ocp-cluster-version-upgrades: Restrict channels to previous-version clusters
Restrict the channels table to clusters running the previous version,
so release-admins who are considering tagging A->B into stable-4.2 (or
whatever) can see how many folks are running A in stable-4.2, because
those are the folks who will see the new edge.  We don't really care
how many folks are running stable-4.2 on other releases.