Add OCM gateway configurations
Currently the configuration of the OCM gateway is part of the container
image. This means that changing the configuration requires building a
new version of the image, and pushing it to the corresponding
environment. This patch creates configmaps in the `app-interface`
project containing the same configuration. The API gateway will then be
changed to use these configmaps, and then it will be possible to change
the configuration without having to build new container images.

Related: https://gitlab.cee.redhat.com/service/uhc-gateway/merge_requests/141
Signed-off-by: Juan Hernandez <jhernand@redhat.com>