datahub-psi: add source for ppc64
People are asking about Power clusters as well. For some reason,
virt-what (the tool that gives us the virt platform data) doesn't offer
a generic Power fact like it does with System Z. As a result, we have to
use a regular expression when finding all of the Power variants.