Add Shannon Hughes to Insights clusters
Shannon is a support lead for Insights, so we are creating a new role
for him that grants him access to our clusters, which will allow him to
use our cluster's OAuth to log in to our Kibana and Grafana services.

At this time we don't intend to grant access to any namespaces on
our clusters.