[APPSRE-1434] Fix reports on integration failures
The `set -exvo pipefail` makes the script interrupt the execution on the
first failed command. When the `manual_reconcile.sh` command fails, the
next command, the `gen_report.py`, is not executed.

This patch makes the script go ahead even if the `manual_reconcile.sh`
fails, while still colecting its exit code when it's not 0.

Signed-off-by: Amador Pahim <apahim@redhat.com>