Split redis out from the rest of rbac saas file
Redis is pulled from Dockerhub and does not require auth.  In fact, if
you specify creds, the code that pulls the image to examine its manifest
fails with an auth error.

The real fix is to probably try dockerhub URLs without auth if they fail
with auth.  Another possible fix is to allow authorization overrides in
indivdual resourceTemplates.

In the mean time, let's just split out the redis deployment into a
separate saas file so we can have a different auth config.