Change entitlements-stage to use stage env
The entitlements service uses subscriptions.stage.api.redhat.com, but
this was not accessible from outside the Red Hat internal network.  I
temporarily changed it to subscriptions.qa.api.redhat.com while the IT
team responsible for that service worked on the problem via ticket
INC1278307.

Now that the ticket is resolved, I am changing it back.