Run aws-efs-operator's PR check via `make`
This is a step toward being able to customize/standardize what happens
underneath a PR check. Using a generic `make` target rather than a
script path allows us to use `make`'s chaining etc. rather than
maintaining the logic in a script; and it allows us to change where that
target points without having to come back and update the job.

Long-term, the goal would be to make this the default for all consumers.