Give uhc-portal tests access to 2 new secrets for selenium tests
We want selenium tests to use production SSO, instead of QA SSO which is
unreliable.  With QA SSO is was OK to type a hardcoded username/password
into the login dialog, so the only actual credential we needed was token.
Prod SSO requires actual username & password, added new Vault secrets.

https://issues.redhat.com/browse/SDA-2525