degrade sso alert from Critical to High
The Platform IT team is aware of the current stability issue of
sso.redhat.com, and we have been assured that they are being alerted
every night whenever it flaps.

We have agreed to degrade this alert for the time being until. Note that
we will continue to be alerted if the SSO degradation does affect our services.