Tweak "OCM metrics worker failed" alert to fire continuously
The worker runs once in 8-9 minutes, so increase(errors[5m]) was sometimes 0.
This changes the alert to be continuously true during a long outage.

For last couple days stage infogw-proxy was down, but we didn't get this
alert.
I'm not sure if this was the root cause (it didn't fire because the
intermittent condition never held for 5m?) or it's unrelated issue
with delivery.

https://issues.redhat.com/browse/SDA-2541