Enable external traffic to the gateway for classic traffic
This is dependent on the changes from this e2e PR [1] which sets `APICAST_ALLOW_EXTERNAL_TRAFFIC`
in the deployment config.

Since insights-classic will continue to live in a different cluster (v3 prod), we
this will enable us to allow that traffic with a valid service header and PSK.

Once the flow for insights-classic no longer requires communicating with services
through the gateway, we will be able to pull this configuration out.

[1] https://github.com/RedHatInsights/e2e-deploy/pull/1993/files