Inject the CA bundle into UHC authentication configMap
See https://docs.openshift.com/container-platform/4.5/security/certificates/service-serving-certificate.html#add-service-certificate-configmap_service-serving-certificate

This is necessary for applications on the cluster to be able to
communicate using SSL with other applications that do not expose a
public route. The specific use case this is being added for is to allow
the OCM Service Log application to communicate with the SendGrid
service.