Manage rhsm-ci and rhsm-qa environments
The intent is that `integration-test` is a branch that devs or QE can change as needed
for testing a PR or other change in the CI environment prior to merge.
This is a temporary measure, until emphemeral environments are
implemented.

rhsm-qa will simply be deployed to latest master.

Notably different from stage/prod are the vault secrets for rhsm-db and
host-inventory-readonly-db as these are not managed by app-interface
(yet).