Add Kafka Lag SOP for cloud.redhat.com
I know that there is a lot to update in this SOP in regards to standard
app-sre operating procedures, but I just wanted to get this copied over
to app-interface today and then update it here going foward.

I wanted to link to this from a new alert going in.