Stop trying to deploy aws-efs-operator
This is a community operator, not an OSD operator, so we shouldn't be
trying to deploy it to clusters. This commit removes its saas file, hive
namespaces (we were missing the v4s anyway), and OLM registry (don't
know (or care) whether this will cause the quay repo for the registry to
be deleted).