hive alerts: refine alerts to only page app-sre on high POST counts
Refine alerts so app-sre only gets paged for the only super-dangerous
condition -- a probable infinite loop of POST requests to etcd, local
(hive) or remote (customer).

This is the only situation where we risk filling etcd and destroying
the hive cluster or a customer cluster.

PUT, DELETE, GET, etc do not present a risk to the hive cluster's
or customer cluster's health. Don't page for these.

Refine the alerts to match recent changes to the hive-metrics grafana.

- Take the 5m average and alert if those go over 15 for 1 minute.
  client-go rate limits at 20, so in an infinite loop, we'll see requests
  pegged in the 19-20 range. Normal every-two-hour syncset reconciles
  top out in the 7-8 range. So, anything above 15 for a minute is most
  likely a hotloop.