Add swatch admin endpoints to turnpike routes
The intent is to allow access to hawtio admin console, which lives on
the app at /actuator/hawtio. Note that we've got a couple of PRs in
progress to actually enable these endpoints:

- https://github.com/RedHatInsights/rhsm-conduit/pull/162
- https://github.com/RedHatInsights/rhsm-subscriptions/pull/175

There are a couple of other endpoints we can hit at /actuator to verify
before they land though: namely /actuator/health, /actuator/info, and
/actuator/prometheus.

The hawtio console replaces the functionality that we lost moving from
OpenShift 3 to OpenShift 4, specifically the Java console.

Note that we can be flexible on the prefix as needed.