Add RHSM API url to rhsm-conduit
The pinhead service that we are currently using has been deprecated. The
API has been ported to a different service, but this service has a
slightly different API, and a different endpoint.

Once pinhead is actually decommissioned, we should remove the
PINHEAD_URL parameter, but would like to leave it in place until then in
case we need to roll back this change.

Related:
- https://issues.redhat.com/browse/ENT-2752
- https://github.com/RedHatInsights/rhsm-conduit/pull/161