Drop Hive RemoteKubeClientRequestsHigh alert.
This alert was added to match the one we do for local kube client
requests (to the hive cluster) after a hive bug went into an infinite
loop creating resources with a random name and filled up etcd, bringing
down a cluster.

However this remote alert does not recognize that these requests are
global to all servers, as such this will fire any time the syncset
controller is busy working on a lot of clusters, and is going off
repeatedly during scale testing when things are in fact fine and working
well.

We do not have the same concerns with filling up a remote cluster etcd
for syncset processing as we never generate random names for resources
being synced, worst case here a bug would cause a hotloop syncing the
same resources over and over, but this can be spotted in other ways.
(similar for the machineset controller)

As such I believe it's safe to remove this alert, it was not added for a
real world problem that's been encountered before, and will fire during
normal happy path work.