Move OCM integration _sandbox_ to _OpenShift_ v4
This patch changes the OCM `SANDBOX_...` environment variables to point to the
_OpenShift_ v4 cluster. This means that the OCM integration tests will create
the `sandbox-...` namespaces in that v4 cluster. But the pods that will run
inside those namespaces still connect to the OCM services in the old v3
integration environment. For example, the clusters management service that will
run inside those `sandbox-...` namespaces will still connect to the accounts
management service that runs in the v3 integration environment. That will be
changed in later patches.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>