Add inital SOP for Clowder
This is more a general guide for using Clowder than it is a set of
instructions for specific scenarios that may be encountered during an
outage or otherwise.

I am open to reformatting the SOP into something else if needed by
AppSRE.  This is just a first pass, and we haven't had any real outages
in production to collect specific use cases.