Promote saas-managed-cluster-config 2951ed48aa231221e76f0cc5b83827a71d58d746
https://github.com/openshift/managed-cluster-config/compare/36143a3edbbec9dcffd7c765f3db02859710e5da...2951ed48aa231221e76f0cc5b83827a71d58d746

---
```
commit: f18f70096b3be13c08287d30efd87b10aba591c9
author: Naveen Malik
Quick fixes for broken CSSRE rbac and a missing NS for SA cronjob

The clusterrole for CSSRE was renamed with backplane work.  You cannot change roleRef in an clusterrolebinding, so this broken existing cluster RBAC.  The fix is simply rename the clusterrolebinding which will delete the old one and create the new one.  It will use a unique name.

The namespace used by SA cronjob was removed.  Added it back to a SSS w/ Upsert mode so it can safely be deleted in the future.

```