Bump the stage image tag to trigger a rollout
We had an internal base URL change required for a service, which requires the
per-pod cache be cleared, which only happens on boot.

In order to handle this in app-interface, we've created a new tag for `96fa24a`
as `96fa24a-121820` which is inline with `96fa24a` but should trigger a new
deployment.