Add egress to OCM clusters service Envoy
This patch adds to the OCM clusters management service Envoy a new egress
listener that will be used to handle traffic destinated to other services. This
traffic will be sent directly to the other services without going via the
_OpenShift_ router or the API gateway.

Note that this change will not have effect inmediately because the backend is
still configured with the old `--gateway-url` option. A separate patch will
change that once we have verified that this configuration works correctly.

Related: https://issues.redhat.com/browse/SDA-3274
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>