Disable TLS for comminication between Envoy and OCM CS
This patch changes the configuration of the Envoy proxy of the clusters
management service of the production environment so that the communication with
the backend will use plain HTTP (over Unix sockets) instead of HTTPS.

This only affects the production environment, and only the alternative route
for traffic that goes through Envoy, which is not yet in use. So the only
impact in the production environment is that the pods will be restarted.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>