Remove `-envoy` suffix from integration OCM CS route
This patch renames the OCM CS route of the integration environment from
`clusters-mgmt-envoy` to `clusters-mgmt`.

The reason to do this is that the strategy that we used in the integration and
staging environments to switch traffic didn't work well. We created the new
`clusters-mgmt-envoy` route using an alternative host name, and then in order
to do the switch we updated both the existing `clusters-mgmt` and
`clusters-mgmt-envoy` route swapping the host names. That resulted in both
routes temporarely claiming the same host name, and therefore the change failed
and required manual intervention. To avoid that in the production environment
we will instead do the switch changing the target service of the existing
`clusters-mgmt` route, without touching the hosts. That will result in a
discrepancy in the names of the active routes: it would be `clusters-mgmt` in
production but `clusters-mgmt-envoy` in integration and stage. Avoiding that
discrepancy is the reason for this patch.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>