Configure OCM Envoy to use remote address
Envoy tries to extract the client address from the `x-forwarded-for` header and
from the actual remote address. It then uses that information decide if the
requests are _edge_ or not. Edge requests are the ones arriving directly from
an untrusted client, and are more sanitized than other requests. In particular
the `x-request-id` header is removed from edge requests but not from internal
requests. In our case Envoy is considering that requests from the OpenShift
router aren't edge, because the IP address used by the router is inside one of
the RFC1918 private ranges. The net result is that if a client sends the
`x-request-id` header to the OpenShift router, Envoy will accept it as if it
was an internal request, write it to the logs, and forward it to the
application without change. This patch changes the Envoy configuration so that
it will classify those requests as edge, and therefore it will remove the
`x-request-id` header sent by the user and replace it with a randomly generated
one.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>