Use Unix sockets for OCM CS egress
This patch changes the Envoy listener used for the outbound traffic of the OCM
clusters management service so that it uses a Unix socket instead of a TCP
port.

This was the intended configuration, but there was a mistake in the previous
patch that introduced the egress listener.

Note that this will have no effect in the production environments (other than
restarting the pods) because they are still configured with the `--gateway-url`
pointing to the API gateway of the environment.

Related: https://issues.redhat.com/browse/SDA-3274
Related: https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/14478
Related: https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/14391
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>