MGDSTRM-1225 Add RH-SSO CI/CD definitions
This is based on strimzi-operator and ocm-sendgrid-service. Jenkins jobs
for building all of RH-SSO, and making the operator available are
defined. SaaS file is excluded to avoid scope overlap with another
issue.

The PR check job and script currently do nothing, as we don't have
enough defined in terms of where the source code is going to be.
Therefore, knowing where and how to check the PRs isn't possible at this
point.

I tried to get away with defining as little as possible, to keep what
I'm adding confined to the scope of MGDSTRM-1225.