Promote saas-managed-cluster-config 23d53eb9f31e4fca71ca869a7a6da5e4b0fdd8c4
https://github.com/openshift/managed-cluster-config/compare/f34a299a9905a484499b1fa159d238b9570421ac...23d53eb9f31e4fca71ca869a7a6da5e4b0fdd8c4

---
```
commit: 3648bda79d08c340209358d35882ec4cb8590dfe
author: Taylor Fahlman
Make sure prod has a cluster-monitoring-config cm

commit: 1230f3093e81173a2df28ee8c55fc108d832027f
author: Dominic Finn
none should upgrade OSD like this

commit: 22577c5d80791d67025ebd0da68b04737fe320a5
author: Eric Fried
osd-cluster-ready: restartPolicy: OnFailure

Change the `restartPolicy` for osd-cluster-ready to `OnFailure` so that
it will restart if errors occur. Remove the overrides for `backoffLimit`
and `activeDeadlineSeconds` to let the defaults for those take over.

(Note that this is not really related to the maximum cluster age
observed by the job. The job only fails if an error occurs, not if
health checks are unsuccessful. They tie together only in the sense
that, if the job restarts after one or more failures, it still uses the
inception of the cluster (not the job) to determine the maximum age.)

```