Overhaul OCM clusters service dashboard
This patch changes the OCM clusters management service so that it uses
the Envoy metrics where possible, as the metrics provided by the service
itself are less reliable due to how they are generated. For example,
currently there aren't metrics for the outbound requests sent to the
`service_logs` service because the code doesn't use correctly the OCM
SDK. Using the Envoy metrics solves this problem.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>