Mirror `docker.io/golangci/golangci-lint` to Quay
We're currently using a fixed old golangci-lint v1.10.2 binary installed on Jenkins.
(https://gitlab.cee.redhat.com/app-sre/infra/blob/master/ansible/playbooks/roles/jenkins-slave-service-development/tasks/golangci-lint.yml)
And developers on the SD-A/B teams downloaded the same version.

We don't keep it up to date, because several repos use it — hard to know none of them will break. Plus, it's a hassle for the team to download matching new version (especially if not same between repos).

Running from a container is an easy way to allow each repo to pick its own version, which will allow independent upgrading, with CI confirming it works for this repo, of golangci-lint version, with no coupling to what Jenkins has installed, and no hassle to the team:
https://gitlab.cee.redhat.com/service/uhc-clusters-service/-/merge_requests/2676
But first I need it mirrored, because docker.io is blocked in Jenkins.