Remove unused OCM `api-connection` secret key
The `api-connection` secret key was introduced some time ago with the
intention to use it to tell the OCM services to talk directly to other
OCM sevices without going via the OpenShift router. Since them the
approach has changed and it is now based on Envoy configuration files.
This secret key was never actually used and isn't needed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>