Add `FORCE_MIGRATION` variable to OCM CS saasherder configuration
This patch adds a new `FORCE_MIGRATION` variable to the OCM clusters
management service _saasherder_ configuration files. This variable will
be used when a migration fails and it is necessary to clear the database
dirty flag and force a migration version number. For example, if the
last successful migration was `123` and a push to production was done
that failed to run migration `124` the database will have the dirty flag
set to `t` an the version to `124`. To recover from that situation it
will be necessary to revert to the previous veersion of the service
*and* set this variable to `123`. After making this change and
recovering the service the variable should be changed again to set it
empty before the next push.

Related: https://issues.redhat.com/browse/SDA-3750
Related: https://gitlab.cee.redhat.com/service/uhc-clusters-service/-/merge_requests/2688
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>