Add CI env vars for new OCM sandbox
The original idea was to switch from old to new cluster
in Vault and/or app-interface.
But that way the change is not immediately tested in CI,
and CI success/failures are not traceable to git history.

IMO the best place to switch clusteres is
uhc-clusters-service repo, where it gets tested.