Test ConfigMap sync to v3 CI
This sets up two new resource templates, one for each ConfigMap in CI (permissions/roles).
Currently this should pull from a test `configmaps` branch, which is in sync
with what's in CI currently.

Assuming this works, we'll namespace the RBAC config repo by environment, and
setup a pair of resourceTemplates per environment, to reference the correct path.