Update to Limitador 0.4.0 to add `name` label
This patch updates to version `0.4.0` of Limitador in order to be able
to add the limit name as a label in the Prometheus metrics.

This only affects the stage environment. The production environment will
be updated in a separate change.

Related: https://gitlab.cee.redhat.com/service/rate-limiting-templates/-/merge_requests/15
Related: https://github.com/3scale-labs/limitador/issues/22
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>