Send OCM integration and stage CS->AMS traffic Envoy
Currently the traffic that originates in the OCM clusters management
service and that goes to the OCM accounts management service doesn't
pass through the Envoy proxy that is in front of the accounts management
service. This patch changes that.

This only affects the integration and stage environments. The production
environment will be changed in a separate patch.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>