Fix label of OCM gateway jobs
In a previous patch I changed the label of the gatewway jobs from
`sd-uhc` to `rhel8` believing that that was required to run the jobs in
the new RHEL8 slave. But turns out that the label is only used to
generate the name of the job, something like `[rhel8] ...`, and that
name is then used to decide in which view to put the job in the UI. The
net result of that change is that the job no longer appears in the
`sd-uhc` view in the UI. This patch fixes that.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>