Add the pr_check job for MAS SSO
JIRA Link: https://issues.redhat.com/browse/MGDSTRM-2241

This is a part of the effort to add the missing pieces to get
MAS SSO to staging. pr_check job has been added to target a
pr_check script in our repo. This would validate the
the existence of the registry images that are about to be synced.