Send OCM production CS->AMS traffic via Envoy
Currently the traffic that originates in the OCM clusters management
service and that goes to the OCM accounts management service doesn't
pass through the Envoy proxy that is in front of the accounts management
service. This patch changes that.

This affects the production environments, but there will be no downtime.
The clusters management service will be restarted and the new pods will
send traffic via the `uhc-acct-mngr-envoy`service while the old pods
will still use the old `uhc-acct-mngr` service, as both are currently
enabled in the accounts management service.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>