Scale OCM gateway to 3 pods
In a previous change it was scaled to 4 pods in an attempt to share load
between pods and reduce the use of CPU in each pod, which was well over
the limits. But apparently that high use of CPU was not caused by real
activity, because it decreased to normal levels when the pods were
restarted.

This patch scales back to 3 pods again, so if/when the problem happens
again we can investigate it in the same initial conditions.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>