Send OCM integration and stage CS->OSL traffic via Envoy
Currently the traffic from the OCM clusters management service to the
OCM service logs doesn't go through the Envoy proxy that is in front of
the OCM service logs. This patch changes that.

This only affects the integration and stage environments. The production
environment will be updated in a different patch.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>