Force OCM migration in stage and integration
Migration `1616431696_update_new_machine_type_size.up.sql` failed with
the following error:

```
2021-03-29T17:38:35Z [ERROR] [initialize/cmd.go:114] [1b288145-d976-4235-b942-b59db1657182]-
failed to run migrations: failed to update the database schema: migration failed:
column "options" contains null values in line 0:
ALTER TABLE addon_parameters ADD COLUMN IF NOT EXISTS options JSONB NOT NULL;
```

This patch changes the configuration of the service in the stage and
integration environments so that it will force the immediately previous
migration number and clear the dirty flag.

This will need to be merged *after* the actual fix of the migration, and
will have to be undone inmmediately after the service is restored.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>