Add OCM production clusters management service configmap
This patch adds a new `clusters-service` configmap to the OCM
namespaces. This configmap will initially have only a
`install-config.yaml` key containing the default OpenShift installer
configuration. This configuration will be merged with the dynamically
generated one, and the result will be used to install clusters.

The intent of this is to provide a mechanism to set alternative image
content sources, for example:

```
install-config.yaml: |
  apiVersion: v1
  kind: InstallConfig
  imageContentSources:
  - source: quay.io/openshift-release-dev/ocp-release
    mirrors:
    - quay.io/openshift-release-dev/ocp-release
    - 950916221866.dkr.ecr.us-east-1.amazonaws.com/ocp-release
  - source: quay.io/openshift-release-dev/ocp-v4.0-art-dev
    mirrors:
    - quay.io/openshift-release-dev/ocp-v4.0-art-dev
    - 950916221866.dkr.ecr.us-east-1.amazonaws.com/ocp-art-dev
```

The initial content is empty, so that it will have no effect in the
production environment till the configmap is changed with the mirrors
required in the production environment.

Note that have no effect till the corresponding change in the OCM
clusters management service has been merged and pushed to the relevant
environments.

Note only affects the production environment. The other environments
have been changed in previous patches.

Related: https://gitlab.cee.redhat.com/service/uhc-clusters-service/-/merge_requests/2742
Related: https://issues.redhat.com/browse/SDA-3109
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>