ruamel roundtrip all apps
We need to add some fields to all the app files. We want to do this
programmatically, there we want to process all the app files with
ruamel with no changes to the content, before using ruamel to modify the
content in the files.

Context: https://issues.redhat.com/browse/APPSRE-2880