Move OCM API tests to SDB worker node
This patch move the OCM API tests jobs to the `sd-service-development-b`
Jenkins worker node. These jobs were already running in that node or in
the `sd-service-development` node because they are labeled with the
`sd-uhc` label. But the `sd-service-development` node is going to be
retired and replaced with a new `rhel8` node.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>