OCM: Bump CS to 2297689f
This contains a single change to remove `install_config_default`
organization capability.

Currently the use of the default OpenShift installer configuration from
the file given in the `--install-config` command line option is limited
to organizations that have the `install_config_default` capability. This
patch removes that capability and makes the use of that default
installer configuration mandatory for all users and organizations.

Related: https://issues.redhat.com/browse/SDA-3946