Add Quay pull credentials to OCM portal `pr_check` job
In order to run multiple copies of the the `pr_check` job of the OCM
portal in the same Jenkins node simultaneously we are going to use a
podman pod that contains several containers, including a private Nginx
image from Quay. To be able to pull that private image we need the
`QUAY_USER` and `QUAY_TOKEN` environment variables populated with the
corresponding values from Vault.

Related: https://gitlab.cee.redhat.com/service/uhc-portal/-/merge_requests/2410
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>