Add _pod_ variable to OCM CS dashboard
This patch adds a new _pod_ variable to the OCM clusters management
service dashboard. The variable will be populated with the names of the
pods, using this Graphana query:

```
query_result(
  count_over_time(
    kube_pod_info{namespace="$namespace",pod=~"^clusters-service-.*$"}[$__range]
  )
)
```

The actual pod name is extracted from the results of that query using
this regular expression:

```
/[(,]pod="([^"]*)"/
```

Note that the we the `count_over_time` Prometheus function to be able to
get the names of all the pods that existed in the selected time range,
otherwise we would get only the names of the pods that are currently
alive.

Using this variable the panels have been modified so that it is possible
to view the data for all the pods or for the selected pod. This is
particularly convenient for the panel that shows the use of memory,
because it contains already three metrics per pod, and it isn't easy to
inspect visually the three metrics that correspond to a particular pod,
which is what we usually want.

In most cases the only change needed was to add `pod=~"$pod"` to the
Prometheus queries.

Related: https://issues.redhat.com/browse/SDA-4089
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>