[APPSRE-3266]Declare OCM-quay RO cluster in us-west-2
It is *much* smaller than the existing ones. The one in us-east-1 is
using 2% CPU and 18 nodes of m5.4xlarge. If I set a goal of 50% CPU
usage to have (plenty of) room for spikes, and the smallest available
instance in OCM is m5.xlarge, I need 4 to 5 nodes.

We also grant app-sre access to this cluster as part of its normal
role.

TODO: Define VPC and peerings, but that seems to be manual.