Enable concurrent builds of OCM portal
The `pr_check.sh` script of the OCM portal has been recently changed to
run the Selenium tests in a way that supports multiple simultaneous
executions in the same Jenkins worker node. This patch changes the
configuration of the job to enable that.

Related: https://gitlab.cee.redhat.com/service/uhc-portal/-/merge_requests/2410
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>