Swatch: bump worker threads back up to 12
Given a nightly run hadn't happened successfully yet, I decided to take
a quick peek and it's not executing quickly enough to finish in the
alotted time.

See
https://grafana.app-sre.devshift.net/d/MMzCEd9Mk/subscription-watch?editPanel=78&viewPanel=78&orgId=1&from=now-6h&to=now

!19336 reduced threads to 2 due to memory issues.

Essentially rate appears to be ~2200 messages / hour... Given ~75k
messages, it's not going to complete in 24 hours (which is our goal).
Given that 24 was causing some sort of memory issue, I'm going to try
something in-between. I'll check again in the morning to see how things
worked out.