Bump OCM gateway to c7efb00
The more relevant change in this version is the introduction of the
Envoy proxy that will eventually replace the Apache server. Note that
this is initially connected only to a temporary route tht will be used
to test it; real traffic will still go via the Apache server. We will
later switch the traffic changing the target service of the
_api.openshift.com_ route.

Related: https://issues.redhat.com/browse/SDA-4220
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>