Add Envoy panels to OCM gateway dashboard
This patch adds to the OCM gateway dashboard new panels for the Envoy
proxy that will eventually replace the Apache server.

The new panels are mostly equivalent to the old ones, but using the
metrics provided directly by the Envoy proxy instead of the metrics
provided by the Apache exporter that we currently use.

The existing panels are preserved and will be removed when the Apache
server itself is removed.

Related: https://issues.redhat.com/browse/SDA-4220
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>