Set yupana memory request to 11Gi
Yupana has a memory leak and is constantly getting OOMKilled.  The
impact of this is made substantially worse because the k8s scheduler
does not leave enough memory for Yupana because its memory request is so
low (1Gi).  This causes the node it gets scheduled on to overcommit its
memory, causing System OOMs (meaning the OOMKiller gets invoked outside
of a container's cgroup).