Freeze and orphan OpenShift v3 Hive Shards.
Product requirements for Hive now require us to move to v1 CRDs so we
can run on upcoming OpenShift releases.

As such we must orphan the v3 shards which are being migrated off of, so
we can continue development. This change pins hive-stage and
hive-integration to the same commit in production as of yesterday, and
adds warnings on these as well as hive-production not to update the
commit hash.

I will be updating the hive push to prod SOP and notifying team
explicitly not to update these SHA's going forward.